% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R
\name{calculateGUI}
\alias{calculateGUI}
\alias{GUI}
\title{A graphical user interface for the package GFD}
\usage{
calculateGUI()
}
\description{
This function provides a graphical user interface for calculating statistical
tests in general factorial designs.
}
\details{
The function produces a GUI for the calculation of the test statistics and 
for plotting. Data can be loaded via the "load data" button. The formula, 
number of permutations (default: 10,000) and the significance level alpha
(default: 0.05) need to be specified. If the plot option is chosen, an
additional window opens containing information on the plots.
}
