% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{curdies}
\alias{curdies}
\title{Curdies river data set}
\format{A data frame with 36 rows and 3 variables:
\describe{
  \item{season}{a factor with levels "SUMMER" and "WINTER"}
  \item{site}{a factor with levels 1 to 6, nested within "season"}
  \item{dugesia}{number of flatworms counted on a particular stone (in no./dm^2)}
}}
\source{
\url{http://users.monash.edu.au/~murray/AIMS-R-users/ws/ws7.html}
}
\usage{
data(curdies)
}
\description{
A dataset containing the number of flatworms (dugesia) sampled in two seasons at different sites in the Curdies River in Western Victoria.
}
\keyword{datasets}

