% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGFD.R
\name{GFD_GUI}
\alias{GFD_GUI}
\alias{GFDsurvGUI}
\title{A shiny app for the package GFD}
\usage{
GFD_GUI()
}
\description{
This function provides a shiny app for calculating
GFD and QANOVA test statistics and respective p-values.
}
\author{
Philipp Steinhauer
}
