% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{pizza}
\alias{pizza}
\title{Pizza delivery times}
\format{A data frame with 16 rows and 6 variables:
\describe{
  \item{Crust}{a factor with levels "thick" and "thin"}
  \item{Coke}{whether or not Coke was ordered with the pizza ("yes" or "no")}
  \item{Bread}{whether or not garlic bread was ordered with the pizza ("yes" or "no")}
  \item{Driver}{the sex of the driver, a factor with levels "M" and "F"}
  \item{Hour}{time of order in hours after midnight}
  \item{Delivery}{Delivery time in minutes}
}}
\source{
\url{http://www.statsci.org/data/oz/pizza.html}
}
\usage{
data(pizza)
}
\description{
A dataset containing the delivery times for pizza (in minutes) under different conditions.
}
\keyword{datasets}
