% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{startup}
\alias{startup}
\title{Startup Costs of five different companies}
\format{A data frame with 60 rows and 2 variables:
\describe{
  \item{Costs}{price, in thousands of dollars}
  \item{company}{company, a factor with levels "pets", "pizza", "gifts", "shoes" and "bakery"}
}}
\source{
\url{https://college.cengage.com/mathematics/brase/understandable_statistics/7e/students/datasets/owan/frames/frame.html}
}
\usage{
data(startup)
}
\description{
A dataset containing the startup costs (in thousands of dollars) of five companies.
}
\keyword{datasets}
